elided SAC support for depricated system functions and got rid of all warnings on OSX

+++ src/system/src/GetOpt/getopt.c
@@ -154,7 +154,7 @@ int getopt_sac(const char *opts)
     }
     else
     {
-      if (argind >= strlen(arg))
+      if ((size_t)argind >= strlen(arg))
       {
         optopt = '\0';
       }
